chore: removed repeated running of live server for each tests
