--- e2e/babel_testEnvironment.js
@@ -1,4 +1,0 @@
-require('@babel/register')({
-  extensions: ['.ts', '.js']
-})
-module.exports = require('./testEnvironment.js')
