+++ e2e/globalTeardown.js
@@ -1,0 +1,3 @@
+module.exports = async () => {
+  global.__LIVESERVER__.kill()
+}
